requirements.txt
Pillow==2.8.1
requests==2.7.0
